# swagger/api-gateway.yaml
openapi: 3.0.3
info:
  title: SmartHouse API Gateway
  version: 1.0.0
  description: |
    API-шлюз системы SmartHouse
...
security:
  - BearerAuth: []
paths:
...
  /houses/{houseId}/devices:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    get:
      tags: [Devices]
      summary: Список устройств в доме
      responses:
        '200':
          description: Устройства в доме
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Devices]
      summary: Регистрация нового устройства в доме
      description: |
        Регистрация нового устройства в доме
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegisterRequest'
      responses:
        '201':
          description: Устройство зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /houses/{houseId}/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/HouseId'
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Devices]
      summary: Получить информацию об устройстве
      responses:
        '200':
          description: Устройство
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Devices]
      summary: Обновить устройство
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: Обновлённое устройство
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Devices]
      summary: Удалить устройство из дома
      responses:
        '204':
          description: Устройство удалено
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /houses/{houseId}/devices/{deviceId}/telemetry:
    parameters:
      - $ref: '#/components/parameters/HouseId'
      - $ref: '#/components/parameters/DeviceId'
      - in: query
        name: metricKey
        schema:
          type: string
        description: Ключ метрики, например "temperature", "power", "state"
      - in: query
        name: from
        schema:
          type: string
          format: date-time
        description: Начало интервала
      - in: query
        name: to
        schema:
          type: string
          format: date-time
        description: Конец интервала
      - in: query
        name: limit
        schema:
          type: integer
        description: Максимальное количество записей
      - in: query
        name: sort
        schema:
          type: string
          enum: [asc, desc]
          default: desc
        description: Направление сортировки
    get:
      tags: [Telemetry]
      summary: Получить историю телеметрии устройства
      responses:
        '200':
          description: Список записей телеметрии
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryPoint'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    HouseId:
      name: houseId
      in: path
      required: true
      schema:
        type: string
      description: Идентификатор дома
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
      description: Идентификатор устройства
...
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
        serialNumber:
          type: string
        deviceModelId:
          type: string
        houseId:
          type: string
        name:
          type: string
        status:
          type: string
          description: Статус устройства
      required:
        - id
        - serialNumber
        - deviceModelId
        - houseId
        - name
        - status
    DeviceRegisterRequest:
      type: object
      properties:
        serialNumber:
          type: string
        deviceModelId:
          type: string
        name:
          type: string
        initialAttributes:
          type: array
          items:
            $ref: '#/components/schemas/DeviceAttribute'
      required:
        - serialNumber
        - deviceModelId
        - name
    DeviceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
    DeviceAttribute:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
...
    TelemetryPoint:
      type: object
      properties:
        deviceId:
          type: string
        houseId:
          type: string
        metricKey:
          type: string
        value:
          type: number
        unit:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - deviceId
        - houseId
        - metricKey
        - value
        - timestamp
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required:
        - code
        - message